# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Leadership Engagement Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{id}/invitations:
    get:
      operationId: listEventInvitations
      tags: [invitations]
      summary: List all invitations for an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invitation"
    post:
      operationId: createInvitation
      tags: [invitations]
      summary: Invite a person to an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationInput"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
  /events/{id}/invitations/bulk:
    post:
      operationId: bulkCreateInvitations
      tags: [invitations]
      summary: Invite multiple people to an event in a single transaction
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkInvitationInput"
      responses:
        "201":
          description: Bulk invitations result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkInvitationResult"
    patch:
      operationId: bulkUpdateInvitations
      tags: [invitations]
      summary: Update attendance status for multiple invitations in one call
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkInvitationStatusInput"
      responses:
        "200":
          description: Bulk update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkInvitationStatusResult"
  /invitations/{id}:
    patch:
      operationId: updateInvitation
      tags: [invitations]
      summary: Update attendance status for an invitation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
        "404":
          description: Not found
    delete:
      operationId: deleteInvitation
      tags: [invitations]
      summary: Remove an invitation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Virtual Meetings ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    Person:
      type: object
      required: [id, name, email, role, homeCity, homeState, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        title:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        role:
          type: string
          enum: [executive, secondary_leader, staff]
        homeCity:
          type: string
        homeState:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Event:
      type: object
      required: [id, name, location, city, state, startDate, eventType, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: string
        city:
          type: string
        state:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        eventType:
          type: string
          enum: [summit, conference, marketing, leadership, regional, other]
        leaderCount:
          type: integer
        inviteeCount:
          type: integer
        attendeeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Invitation:
      type: object
      required: [id, eventId, personId, status, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        personId:
          type: integer
        status:
          type: string
          enum: [invited, attended, no_show, declined]
        notes:
          type: ["string", "null"]
        graphEventId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        person:
          $ref: "#/components/schemas/Person"
        event:
          $ref: "#/components/schemas/Event"
    InvitationInput:
      type: object
      required: [personId]
      properties:
        personId:
          type: integer
        notes:
          type: string
        createCalendarEvent:
          type: boolean
          description: "If true (default), create an Outlook calendar event via Microsoft Graph for this invitation."
    InvitationUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [invited, attended, no_show, declined]
        notes:
          type: string
    BulkInvitationInput:
      type: object
      required: [personIds]
      properties:
        personIds:
          type: array
          items:
            type: integer
          minItems: 1
        createCalendarEvent:
          type: boolean
          description: "If true (default), create Outlook calendar events for each new invitation."
    BulkInvitationResult:
      type: object
      required: [created, skipped, invitations]
      properties:
        created:
          type: integer
          description: Number of new invitations created.
        skipped:
          type: integer
          description: Number of people already invited (skipped, not re-invited).
        invitations:
          type: array
          items:
            $ref: "#/components/schemas/Invitation"
    BulkInvitationStatusInput:
      type: object
      required: [updates]
      properties:
        updates:
          type: array
          items:
            type: object
            required: [id, status]
            properties:
              id:
                type: integer
              status:
                type: string
                enum: [invited, attended, no_show, declined]
          minItems: 1
    BulkInvitationStatusResult:
      type: object
      required: [updated]
      properties:
        updated:
          type: integer
          description: Number of invitation statuses updated.

    # ─── Virtual Meeting ─────────────────────────────────────────────────────
...